requirements.txt
urllib3==1.26.6
requests
DrissionPage==v4.0.4.23
# undetected-chromedriver
sanic==23.12.1
sanic-ext
sanic-openapi
websockets==10.0
pydub==0.25.1
SpeechRecognition==3.10.4
# pillow==10.4.0
# pytesseract==0.3.13
# opencv-python==4.10.0.84
openai-whisper==20231117
transformers==4.40.0
curl_cffi==0.7.1